remove types file on ChartMenu
